Add regular item use handler. Adjust buff source to int32.

+++ atlas.com/channel/skill/handler/beginner_recovery.go
@@ -25,7 +25,7 @@ func UseBeginnerRecovery(l logrus.FieldLogger) func(ctx context.Context) func(m 
 				_ = skill.ApplyCooldown(l)(ctx)(m, info.SkillId(), effect.Cooldown())(characterId)
 			}
 
-			_ = buff.Apply(l)(ctx)(m, characterId, info.SkillId(), effect.Duration(), effect.StatUps())(characterId)
+			_ = buff.Apply(l)(ctx)(m, characterId, int32(info.SkillId()), effect.Duration(), effect.StatUps())(characterId)
 			return nil
 		}
 	}
